+++ engines/dcnn.h
@@ -1,0 +1,6 @@
+#ifndef PACHI_ENGINES_DCNN_H
+#define PACHI_ENGINES_DCNN_H
+
+struct engine *engine_dcnn_init(char *arg, struct board *b);
+
+#endif /* PACHI_ENGINES_DCNN_H */
