+++ pachi.c
@@ -16,6 +16,7 @@
 #include "engines/patternscan.h"
 #include "engines/patternplay.h"
 #include "engines/joseki.h"
+#include "engines/dcnn.h"
 #include "t-unit/test.h"
 #include "uct/uct.h"
 #include "distributed/distributed.h"
@@ -81,8 +82,8 @@ static void usage(char *name)
 		"  -c, --chatfile FILE               set kgs chatfile \n"
 		"  -d, --debug-level LEVEL           set debug level \n"
 		"  -D                                don't log board diagrams \n"
-		"  -e, --engine ENGINE               select engine: (default uct) \n"
-		"                                    random|replay|montecarlo|uct|distributed|dcnn|patternplay \n"
+		"  -e, --engine ENGINE               select engine (default uct). Supported engines: \n"
+		"                                    uct, dcnn, patternplay, replay, random, montecarlo, distributed \n"
 		"  -f, --fbook FBOOKFILE             use opening book \n"
 		"  -g, --gtp-port [HOST:]GTP_PORT    read gtp commands from network instead of stdin. \n"
 		"                                    listen on given port if HOST not given, otherwise \n"
